\name{e.cluster.decision}
\alias{e.cluster.decision}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
cluster detection using echelon
}
\usage{
e.cluster.decision(reg_data, log.lambda)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{reg_data}{
 .
}
  \item{log.lambda}{
 .
}
}
