\name{e.cluster.dendrogram}
\alias{e.cluster.dendrogram}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
draw the detected clusters on the dendrogram
}
\usage{
e.cluster.dendrogram(echelon.obj, n.sim, cluster.legend.pos, cluster_reg, p_rank, para)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{echelon.obj}{
.
}
  \item{n.sim}{
.
}
  \item{cluster.legend.pos}{
.
}
  \item{cluster_reg}{
.
}
  \item{p_rank}{
.
}
  \item{para}{
.
}
}
