\name{e.cluster.map}
\alias{e.cluster.map}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
draw the detected clusters on the map
}
\usage{
e.cluster.map(x, c_separates, locs, coo, rin, p_rank, cluster_reg, n.sim, cluster.type)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
.
}
  \item{c_separates}{
.
}
  \item{locs}{
.
}
  \item{coo}{
.
}
  \item{rin}{
.
}
  \item{p_rank}{
.
}
  \item{cluster_reg}{
.
}
  \item{n.sim}{
.
}
  \item{cluster.type}{
.
}
}
