\name{e.dendrogram.axis}
\alias{e.dendrogram.axis}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
axis for echelon dendrogram
}
\usage{
e.dendrogram.axis(main, ylab, yaxes, ylim, xaxes, xdper, dmai, peaks, x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{main}{
.
}
  \item{ylab}{
.
}
  \item{yaxes}{
.
}
  \item{ylim}{
.
}
  \item{xaxes}{
.
}
  \item{xdper}{
.
}
  \item{dmai}{
.
}
  \item{peaks}{
.
}
  \item{x}{
.
}
}
