\name{e.dendrogram}
\alias{e.dendrogram}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
draw the echelon dendrogram
}
\usage{
e.dendrogram(peaks, locs, x, separates, c_separates, number, parents, pare_locs, 
            progeny, symbols, col.symbols, cex.symbols, lwd, col, ens, adj.ens, col.ens, 
            cex.ens, limb)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{peaks}{
.
}
  \item{locs}{
.
}
  \item{x}{
.
}
  \item{separates}{
.
}
  \item{c_separates}{
.
}
  \item{number}{
.
}
  \item{parents}{
.
}
  \item{pare_locs}{
.
}
  \item{progeny}{
.
}
  \item{symbols}{
.
}
  \item{col.symbols}{
.
}
  \item{cex.symbols}{
.
}
  \item{lwd}{
.
}
  \item{col}{
.
}
  \item{ens}{
.
}
  \item{adj.ens}{
.
}
  \item{col.ens}{
.
}
  \item{cex.ens}{
.
}
  \item{limb}{
.
}
}
