\name{e.main}
\alias{e.main}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
main program
}
\usage{
e.main(x, rin, T)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
.
}
  \item{rin}{
.
}
  \item{T}{
.
}
}
