\name{e.monte.bin}
\alias{e.monte.bin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
Monte Carlo testing based on Binomial model
}
\usage{
e.monte.bin(rin,cas,pop,n.sim,K,cluster.type)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rin}{
.
}
  \item{cas}{
.
}
  \item{pop}{
.
}
  \item{n.sim}{
.
}
  \item{K}{
.
}
  \item{cluster.type}{
.
}
}
