\name{e.monte.poi}
\alias{e.monte.poi}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
Monte Carlo testing based on Poisson model
}
\usage{
e.monte.poi(rin, cas, pop, ex, n.sim, K, cluster.type)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rin}{
.
}
  \item{cas}{
.
}
  \item{pop}{
.
}
  \item{ex}{
.
}
  \item{n.sim}{
.
}
  \item{K}{
.
}
  \item{cluster.type}{
.
}
}
