\name{e.profile}
\alias{e.profile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
echelon profiles
}
\usage{
e.profile(peaks, parents, separates)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{peaks}{
.
}
  \item{parents}{
.
}
  \item{separates}{
.
}
}
