\name{e.scan.pop}
\alias{e.scan.pop}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
echelon scan based on a rate of population
}
\usage{
e.scan.pop(x, locs, peaks, c_separates, parents, K, pop)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
.
}
  \item{locs}{
.
}
  \item{peaks}{
.
}
  \item{c_separates}{
.
}
  \item{parents}{
.
}
  \item{K}{
.
}
  \item{pop}{
.
}
}
