\name{e.scan}
\alias{e.scan}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
echelon scan based on the number of regions
}
\usage{
e.scan(x, locs, peaks, c_separates, parents, K)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
.
}
  \item{locs}{
.
}
  \item{peaks}{
.
}
  \item{c_separates}{
.
}
  \item{parents}{
.
}
  \item{K}{
.
}
}
