\name{nlimb}
\alias{nlimb}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Echelon analysis for R
}
\description{
auxiliary of echelon profiles
}
\usage{
nlimb(x, k2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
.
}
  \item{k2}{
.
}
}
